core/standards/web-apps/media/semantics/preload/none/preload-none-event-order-play-after-suspend-mp4.html

<!doctype html>
<!-- This file was generated from 'preload-none-event-order-play-after-suspend.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:none event order when media is played after suspend is fired - mp4</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../support/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="none" controls></audio>
<video preload="none" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:none event order when media is played after suspend is fired - mp4", "HTMLVideoElement.preload:none event order when media is played after suspend is fired - mp4", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
    var tag_name = vars[1];
    var events_actual = '';
    var node = document.getElementsByTagName(tag_name)[0];
    if (!node.canPlayType(tag_name + '/mp4')) {
        t.step(function() {
        assert_unreached("Cannot play '" + tag_name + "/mp4'");
        t.done();
        });
    }
    var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
    for (var e in events) {
        node.addEventListener(events[e], handleEvent, false);
    }
    node.addEventListener('suspend', startTest, false);
    node.src = '/core/standards/web-apps/media/network/range-request-log/range-request.php?rate=100000&fileloc=../../support/preload.mp4&nocache=' + Math.random();

    function startTest(e) {
        node.removeEventListener('suspend', startTest, false);
        node.play();
    }

    function handleEvent(e) {
        if (e.type == 'error') {
        t.step(function() {
            assert_unreached("Error event was fired. Error code: " +  e.target.error.code);
            t.done();
        });
        }
        events_actual += e.type + ' ';
        if (e.type == 'canplaythrough') endTest();
    }

    function endTest() {
        t.step(function() {
        assert_regexp_match(events_actual, /^loadstart suspend (progress )+loadedmetadata (progress )*loadeddata (progress )*canplay (progress )*canplaythrough $/g, 'Event order check');
        assert_equals(node.preload, 'none', "'preload' value should not be affected by playing the media");
        t.done();
        });
    }
    });
});
</script>
